change decimals in formula

+++ backend/app/engine/projects/rewards/allocations/quadratic_funding.py
@@ -8,6 +8,8 @@ from app.engine.projects.rewards import ProjectAllocations
 
 class QuadraticFundingAllocations(ProjectAllocations):
     def _calc_allocations(self, allocations: List[AllocationItem]) -> Decimal:
-        sum_sqrts = sum([sqrt(allocation.amount) for allocation in allocations])
+        sum_sqrts = sum(
+            [Decimal(sqrt(allocation.amount)) for allocation in allocations]
+        )
         quadratic_value = sum_sqrts**2
-        return Decimal(quadratic_value)
+        return quadratic_value
